requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pandas>=2.2.0
numpy>=2.1.0
sqlalchemy>=2.0.30
plotly>=6.0.0
matplotlib>=3.9.0
seaborn>=0.13.0
scikit-learn>=1.6.0
openai>=1.99.1
pydantic>=2.10.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
rich>=13.9.0
streamlit>=1.49.0
requests>=2.32.0
psycopg[binary]>=3.2.0
